Additional CR Feedback and test changes
